Project to add and erase a list of items
On this code I show how is possible to add items through a input element and show them on a list by clicking on a button. Also I developed a second button in which the first hundred items will be erased. Them the user is free to add new items to the list again.