Update Adding to a list and erasing it
On this change I created a code that the use can not enter a number in the first input. The first input is for text only. I also developed a second input in which the user can type (here he must type a number) the quantity of tasks he wants to save. For last I creater a second button that will erase the list when the user click on it.